# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rento Vroom API",
    "version": "1.0.0",
    "description": "REST API for the Rento Vroom website and future apps. Every error response is `{ error: { code, message, fields? } }`. Requests that change data from a browser must come from the website's own origin."
  },
  "servers": [
    {
      "url": "https://api.rentovroom.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rv_access",
        "description": "The 15-minute access token, set by the sign-in routes (the website)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The same access token in an Authorization header (mobile apps)"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code, e.g. VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "description": "A message that can be shown to the user"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "One message per invalid input field"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "TestPayment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The Stripe PaymentIntent id (pi_…)"
          },
          "clientSecret": {
            "type": "string",
            "description": "Lets Stripe.js on the page confirm this one payment. Never logged or stored."
          },
          "amountCents": {
            "type": "integer"
          },
          "currency": {
            "type": "string",
            "enum": [
              "nzd"
            ]
          }
        },
        "required": [
          "id",
          "clientSecret",
          "amountCents",
          "currency"
        ]
      },
      "TestPaymentStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "The PaymentIntent status: succeeded, processing, requires_payment_method, requires_action, canceled, …"
          },
          "amountCents": {
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "paymentMethod": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "type": {
                "type": "string",
                "description": "The payment method type, e.g. card or link"
              },
              "wallet": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "apple_pay or google_pay when a card was paid through a wallet"
              },
              "brand": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "The card brand, e.g. visa"
              },
              "last4": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "type",
              "wallet",
              "brand",
              "last4"
            ],
            "description": "Null until the payment has been attempted"
          },
          "webhookReceived": {
            "type": "boolean",
            "description": "Whether Stripe’s payment_intent.succeeded event reached the webhook"
          }
        },
        "required": [
          "id",
          "status",
          "amountCents",
          "currency",
          "paymentMethod",
          "webhookReceived"
        ]
      },
...
      "UploadTarget": {
        "type": "object",
        "properties": {
          "driver": {
            "type": "string",
            "enum": [
              "local",
              "cloudinary"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "PUT",
              "POST"
            ]
          },
          "url": {
            "type": "string"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "PUT: send the file as the body with these"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "POST: a multipart form with these fields and the file as `file`"
          },
          "maxBytes": {
            "type": "integer"
          }
        },
        "required": [
          "driver",
          "method",
          "url",
          "maxBytes"
        ],
        "description": "Where to send one file. Afterwards, attach it with the key the upload returned: local uploads answer `{ key }`, Cloudinary answers `{ public_id }`."
      },
      "UploadRequest": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "VEHICLE_PHOTO",
              "VEHICLE_DOCUMENT"
            ]
          },
          "vehicleId": {
            "type": "string",
            "pattern": "^[0-9a-f]{24}$"
          },
          "contentType": {
            "type": "string",
            "maxLength": 100
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 15728640
          }
        },
        "required": [
          "purpose",
          "vehicleId",
          "contentType",
          "size"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/payments/test": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Start a NZ$1 test payment in the Stripe sandbox",
        "description": "Admin only. Checks the Stripe keys, Apple Pay, Google Pay and the webhook before the booking flow exists. Refused (409 LIVE_MODE) with live keys, and 503 until the keys are set.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The payment, ready for Stripe.js to confirm",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPayment"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with existing data, e.g. the email address already has an account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A service this needs isn't set up or available yet, e.g. payments before the Stripe keys are set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/payments/test/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "How a test payment went",
        "description": "Admin only. Its status, how it was paid (card, Apple Pay, Google Pay) and whether its webhook arrived.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "The PaymentIntent id (pi_…)"
            },
            "required": true,
            "description": "The PaymentIntent id (pi_…)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The test payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestPaymentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A service this needs isn't set up or available yet, e.g. payments before the Stripe keys are set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/uploads/signature": {
      "post": {
        "tags": [
          "Hosting"
        ],
        "summary": "Where to upload a photo or document",
        "description": "The car’s Host (or staff) only. Photos: JPEG, PNG, WebP or HEIC; documents also PDF; up to 15 MB. 503 UPLOADS_UNAVAILABLE until storage is set up in production.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The upload target",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTarget"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input. `error.fields` has one message per invalid field.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session has ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but this account can't do this (or the request came from an untrusted origin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A service this needs isn't set up or available yet, e.g. payments before the Stripe keys are set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}